+++ Source/BlueprintAssist/Public/BlueprintAssistGlobals.h
@@ -27,3 +27,12 @@ BLUEPRINTASSIST_API DECLARE_LOG_CATEGORY_EXTERN(LogBlueprintAssist, Log, All);
 #else
 #define BA_GET_FONT_STYLE FEditorStyle::GetFontStyle
 #endif
+
+// UE 5.6+ 类型别名：处理 API 变化
+#if BA_UE_VERSION_OR_LATER(5, 6)
+using FBAMetaData = class FMetaData;
+using FBAVector2 = FVector2f;
+#else
+using FBAMetaData = class UMetaData;
+using FBAVector2 = FVector2D;
+#endif